generator: Allow installing snaps from another channel
We should have a way to validate new releases of the generator.

Let's do this by instalilng from the candidate channel on staging. Then
the process can be something like:

  - Push a release commit
  - Promote the snap build to candidate
  - Test it
  - Push the tag, promote the snap build to stable

This commit adds infrastructure to install snaps from a default channel,
or override the channel per-snap. It also
tries to support refreshing if the channel changes.

We then set the default for 'staging' to 'candidate'.